Disable java attacher on windows
